Added indentation feature for ecukes.
Added new step for setting variables.

Corrected test.  Pressing RET is equivalent to Fn-Return (enter), that
is key bound to newline.  It turns out that <return> is the code for
the return key, that is the key bound to C-j (php-auto-fill).

Switching to a buffer in ecukes does not clear it automatically, so
always assume there is something in your buffer if you use a common
name.

The function beginning-of-line-text cannot be trusted.  Added new
function beginning-of-line-non-whitespace to do what is needed.

Replaced all calls to beginning-of-line-text with new function.

Updated ecukes steps, changed all "I go to the beginning of text" with
"I go to the beginning of non-whitespace".

This replacement fixes the semicolon breaking bug.